+++ spring-integration-core/src/main/java/org/springframework/integration/events/package-info.java
@@ -1,4 +1,4 @@
 /**
  * ApplicationEvents generated by the Spring Integration framework.
  */
-package org.springframework.integration.event;
+package org.springframework.integration.events;
