+++ spring-integration-file/src/main/java/org/springframework/integration/file/event/FileIntegrationEvent.java
@@ -16,7 +16,7 @@
 
 package org.springframework.integration.file.event;
 
-import org.springframework.integration.event.IntegrationEvent;
+import org.springframework.integration.events.IntegrationEvent;
 
 /**
  * @author Gary Russell
