+++ spring-integration-ip/src/main/java/org/springframework/integration/ip/event/IpIntegrationEvent.java
@@ -16,7 +16,7 @@
 
 package org.springframework.integration.ip.event;
 
-import org.springframework.integration.event.IntegrationEvent;
+import org.springframework.integration.events.IntegrationEvent;
 
 /**
  * @author Gary Russell
