+++ spring-integration-stomp/src/main/java/org/springframework/integration/stomp/event/StompIntegrationEvent.java
@@ -17,7 +17,7 @@
 package org.springframework.integration.stomp.event;
 
 import org.springframework.context.ApplicationEvent;
-import org.springframework.integration.event.IntegrationEvent;
+import org.springframework.integration.events.IntegrationEvent;
 
 /**
  * Base class for all {@link ApplicationEvent}s generated by the STOMP Adapters.
